Clean code Sign Up process
+ Move 'sign up' gql mutation to it's single file
+ Create useSignUpMutation personal hook
+ Move form onSubmit logic to view component
